requirements.txt
# Core dependencies for the multimodal content pipeline
openai>=1.0.0
python-dotenv>=1.0.0
pydub>=0.25.1
requests>=2.31.0
pillow>=10.0.0
numpy>=1.24.0
gtts>=2.5.0

# LLM fallback: Google Gemini
google-generativeai>=0.8.0

# Optional: For advanced text-to-speech
elevenlabs>=0.2.0

# Optional: For auto-downloading royalty-free YouTube audio
yt-dlp>=2025.1.15

# Scheduling and automation
schedule>=1.2.0
pyyaml>=6.0

# Utilities
tqdm>=4.66.0

# API server
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
